+++ src/Trakx.Data.Common/Interfaces/Index/IComponentDefinition.cs
@@ -23,7 +23,7 @@
         /// <summary>
         /// Number of decimals by which 1 unit of the ERC20 can be divided.
         /// </summary>
-        int Decimals { get; }
+        ushort Decimals { get; }
 
         /// <summary>
         /// The Id of the token on the CoinGeckoApi
